pyproject.toml
[tool.poetry]
name = "iagon-py"
version = "0.1.0"
description = "A Python client for Iagon's decentralized cloud storage platform."
authors = ["Elder Millenial <eldermillenial@protonmail.com>"]
repository = "https://github.com/theledermillenial/iagon-py"
readme = "README.md"
packages = [{ include = "iagon", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
fsspec = "^2023.9.0"
pycardano = { git = "https://github.com/theeldermillenial/pycardano.git", branch = "bugfix/extended-cip8" }

[tool.poetry.plugins."fsspec.specs"]
"iagon" = "iagon.IagonFS"

[tool.poetry.group.dev.dependencies]
nox = "^2023.4.22"
nox-poetry = "^1.0.3"
black = "^23.9.1"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-order = "^1.1.0"
mypy = "^1.6.0"
flake8 = "^6.1.0"
ruff = "^0.0.292"
bump2version = "^1.0.1"
pre-commit = "^3.4.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
